requirements.txt
mock_django==0.6.10
Django==3.2.6
numpy==1.21.1
pandas==1.2.4
djangorestframework==3.12.4
Pillow==8.4.0
python-decouple==3.5
scikit_learn==0.24.2
django-crispy-forms==1.13.0
gunicorn==20.1.0
python_decouple==3.5